pyproject.toml
[project]
name = "tg-bot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiogram>=3.27.0",
    "pocketbase>=0.16.0",
    "python-dotenv>=1.2.2",
]
